rename: stft.py -> signal.py
