+++ python/tvm/topi/__init__.py
@@ -47,7 +47,7 @@ from .scan import *
 from .einsum import *
 from .unique import *
 from .searchsorted import *
-from .stft import *
+from .signal import *
 from . import generic
 from . import nn
 from . import x86
