+++ python/tvm/topi/cuda/__init__.py
@@ -61,4 +61,4 @@ from .sparse_reshape import *
 from .transform import *
 from .unique import *
 from .searchsorted import *
-from .stft import *
+from .signal import *
